tools/requirements.txt
google-generativeai
python-frontmatter
pytz
